+++ CHANGELOG.md
@@ -5,7 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Using SDK with ARKit
 
+### Added
+- AR Tracking listener
+- Model Gesture listener
 
 ## [1.4.0] - 2018-05-30
 ### Changed